+++ src/parser.ts
@@ -5,7 +5,7 @@ import {
   OperationDefinitionNode,
 } from 'graphql';
 
-import invariant = require('invariant');
+const invariant = require('invariant');
 
 export enum DocumentType {
   Query,